fix: months list by year in detailed reports
